Review graphics (#6)
* Update AnimationUtil

- minor update using `const` parameter;
- change warning level 3 to 4

* Update Linear Algebra models

- minor improvement, review also add new `addBone` function to `SimpleDraw`

* Update Graphics

- minor improvement under the Warning lvl 4

* Update Graphics project

- minor improvement since change enable `-Wall`

* Update Graphics

- reviewing and minor improvements.

+++ Framework/Graphics/Inc/Animator.h
@@ -11,7 +11,7 @@ namespace Omega::Graphics
 		void Terminate();
 
 		void ComputeBindPose();
-		void PlayAnimation(int index);
+		void PlayAnimation(size_t index);
 
 		void StopAnimation(bool stop);
 		void SetTime(float time);
@@ -29,7 +29,7 @@ namespace Omega::Graphics
 		float mMaxDuration = 0.0f;
 		float mTimer = 0.0f;
 		float mCurrentTimer = 0.0f;
-		int mClipIndex = 0;
+		size_t mClipIndex = 0;
 		bool mShowSkeleton = false;
 
 		std::vector<Omega::Math::Matrix4> mBoneMatrices;
