+++ Framework/Graphics/Src/MeshBuffer.cpp
@@ -6,7 +6,7 @@
 using namespace Omega;
 using namespace Omega::Graphics;
 
-void MeshBuffer::Initialize(const void* vertices, int vertexSize, int vertexCount, const uint32_t* indices, int indexCount, bool dynamic)
+void MeshBuffer::Initialize(const void* vertices, unsigned int vertexSize, unsigned int vertexCount, const uint32_t* indices, unsigned int indexCount, bool dynamic)
 {
 	mIndiceCount = indexCount;
 	mVertexSize = vertexSize;
@@ -14,10 +14,10 @@ void MeshBuffer::Initialize(const void* vertices, int vertexSize, int vertexCoun
 
 	// size in memory bytes
 	D3D11_BUFFER_DESC bufferDesc{};
-	bufferDesc.ByteWidth = vertexCount * mVertexSize;
+	bufferDesc.ByteWidth = static_cast<size_t>(vertexCount * mVertexSize);
 	bufferDesc.Usage = dynamic ? D3D11_USAGE_DYNAMIC : D3D11_USAGE_DEFAULT;
 	bufferDesc.BindFlags = D3D11_BIND_VERTEX_BUFFER;
-	bufferDesc.CPUAccessFlags = dynamic ? D3D11_CPU_ACCESS_WRITE : 0; // Use CPU
+	bufferDesc.CPUAccessFlags = dynamic ? D3D11_CPU_ACCESS_WRITE : 0u; // Use CPU
 	bufferDesc.MiscFlags = 0;
 	bufferDesc.StructureByteStride = 0;
